adding serializer in delete shopping cart to get deleted one in responce and by userid ffrom query params

+++ shopping_cart/views.py
@@ -7,8 +7,10 @@ from .serializer import ShoppingCartSerializer
 
 class DeleteShoppingCartView(APIView):
     def delete(self, request):
-        user_id = request.data.get('user_id')
+        user_id = request.query_params.get('user_id')
+        if user_id is None:
+            return Response({'error': 'User ID is missing in query parameters'}, status=status.HTTP_400_BAD_REQUEST)
         shopping_cart = get_object_or_404(ShoppingCart, user_id=user_id)
+        serializer = ShoppingCartSerializer(shopping_cart)
         shopping_cart.delete()
-
-        return Response({'message': 'Shopping cart deleted successfully'}, status=status.HTTP_204_NO_CONTENT)
+        return Response({'message': 'Shopping cart deleted successfully', 'deleted_cart': serializer.data}, status=status.HTTP_204_NO_CONTENT)
